Refactor : Renamed defines.hpp to fltypes.hpp
